% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{ad_get_genome}
\alias{ad_get_genome}
\title{Genome info}
\usage{
ad_get_genome(details = FALSE)
}
\arguments{
\item{details}{Logical. If TRUE, return seqinfo object. Default: FALSE.}
}
\value{
GRanges, one range per transcript
}
\description{
Genome info
}
