% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_cds_env}
\alias{EA_cds_env}
\title{DNA sequence (coding) of all the transcripts in this package's transcript set}
\description{
DNA sequence (coding) of all the transcripts in this package's transcript set
}
\details{
Envirnonment, with transcript names as keys, and DNA sequence (as \code{Biostrings::DNAString}) as values.
}
\author{
Dennis Kostka
}
\keyword{data}
