% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_exn_grl}
\alias{EA_exn_grl}
\title{Exons (coding DNA only) of all the transcripts in this package's transcript set}
\description{
Exons (coding DNA only) of all the transcripts in this package's transcript set
}
\details{
Exons of all transcripts, as \code{GenomicRanges::GRangesList}.
}
\author{
Dennis Kostka
}
\keyword{data}
