% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_cds_by_tx}
\alias{get_cds_by_tx}
\title{Coding sequence, stratified by transcript}
\usage{
get_cds_by_tx(txname)
}
\arguments{
\item{txname}{String. Name of transcript}
}
\value{
DNAString. Coding sequence
}
\description{
Coding sequence, stratified by transcript
}
