% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_exns_by_tx}
\alias{get_exns_by_tx}
\title{Exons, stratified by transcript}
\usage{
get_exns_by_tx(txname)
}
\arguments{
\item{txname}{String. Name of transcript}
}
\value{
GRanges. Exons, sorted 5' to 3'
}
\description{
Exons, stratified by transcript
}
