% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{is_ptc_snv}
\alias{is_ptc_snv}
\title{Query stop-making SNVs by transcript}
\usage{
is_ptc_snv(keys, txname)
}
\arguments{
\item{keys}{Character vector. SNVs to be checked.}

\item{txname}{String. Transcript name.}
}
\value{
Logical. For each key, whether it is PTC-generating in txname.
}
\description{
Query stop-making SNVs by transcript
}
